requirements.txt
pydub
SpeechRecognition
pyaudio
pyserial